pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "linux-ai-server"
version = "0.1.0"
description = "API server for full kernel-level Linux control via REST and MCP"
requires-python = ">=3.11"
license = "GPL-2.0-only"
authors = [{ name = "Zaman Huseyinli" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "psutil>=6.0.0",
    "aiosqlite>=0.20.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "paramiko>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "anyio[trio]>=4.0",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
    "bandit>=1.7.0",
]

[project.scripts]
linux-ai-server = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "PT", "SIM", "TCH"]
ignore = [
  "S101",   # assert (used in tests)
  "S603",   # subprocess call — module is used intentionally
  "S607",   # subprocess partial path — relying on PATH is fine here
  "B008",   # FastAPI Depends() in defaults is the canonical pattern
  "S104",   # bind 0.0.0.0 — server is meant to listen on all interfaces
  "S110",   # try/except/pass — used for legit best-effort cleanup
  "S112",   # try/except/continue — same
  "B904",   # raise-without-from — stylistic
  "A001",   # builtin shadowing in vars
  "A002",   # builtin shadowing in args
  "SIM102", # collapsible-if
  "SIM105", # contextlib.suppress preference
  "SIM117", # multiple-with
  "N806",   # non-lowercase var (FastAPI/SQL idioms)
  "TC001",  # typing-only first-party imports
  "TC002",  # typing-only third-party imports
  "TC003",  # typing-only stdlib imports
  "E402",   # module-import-not-at-top — used after sys.path tweaks
  "B007",   # unused-loop-control-variable
  "B905",   # zip-without-strict
  "S324",   # md5 used as non-crypto content hash
  "S507",   # paramiko AutoAddPolicy — deliberate trust-on-first-use
  "S608",   # f-string SQL with internal column/table names (not user input)
  "S202",   # tarfile extractall on backups we created
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S105", "S106", "S107", "S108", "S311", "F841"]
# Inline shell/SQL snippets and compact scoring blocks — wrapping hurts readability
"app/mcp/tools.py" = ["E501"]
"app/api/memory.py" = ["E501", "E701"]
"app/api/deploy.py" = ["E501"]
"app/api/vps.py" = ["E501"]
"app/core/ci_db.py" = ["E501"]
"app/core/devops_agent.py" = ["E501"]
"app/core/task_queue.py" = ["E501"]
# Config/auth literals are setting names, not real secrets
"app/core/config.py" = ["S105", "S108"]
"app/api/auth.py" = ["S105"]
"app/api/backup.py" = ["S108"]
"app/api/projects.py" = ["S108"]
"app/main.py" = ["S108"]
"app/core/terminal_manager.py" = ["S606"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
